+++ src/utils/variables.ts
@@ -17,3 +17,9 @@ export const getVscodeConfigTarget = (activeEditor?: vscode.TextEditor) => {
 
   return vscodeConfigTarget;
 };
+
+export const getConfiguration = async (type: string) => {
+  const config = vscode.workspace.getConfiguration('jinja2-html-enhancer');
+
+  return config.get(type);
+}